feat: change models to be less fuzzy and reduce overhead
