+++ src/main/java/com/github/playerforcehd/gcaptchavalidator/data/ValidationError.java
@@ -26,13 +26,13 @@ package com.github.playerforcehd.gcaptchavalidator.data;
 
 /**
  * Provide a list of all possible error codes that a validation response can contain.
- *
+ * <p>
  * The specific meaning of the error codes can be looked up on Google documentation.
  * Also provides a method to map the plain error codes to the enum.
  *
  * @author Pascal Zarrad
- * @since 3.0.0
  * @see <a href="https://developers.google.com/recaptcha/docs/verify#error_code_reference">Google ReCaptcha Errors</a>
+ * @since 3.0.0
  */
 public enum ValidationError {
     MISSING_INPUT_SECRET("missing-input-secret"),
@@ -41,6 +41,21 @@ public enum ValidationError {
     INVALID_INPUT_RESPONSE("invalid-input-response"),
     BAD_REQUEST("bad-request"),
     TIMEOUT_OR_DUPLICATE("timeout-or-duplicate");
+
+    /**
+     * The error that has been thrown.
+     * Equals the error codes listed in Googles documentation.
+     */
+    private final String plainError;
+
+    /**
+     * Constructor
+     *
+     * @param plainError The original error string that has been thrown
+     */
+    ValidationError(String plainError) {
+        this.plainError = plainError;
+    }
 
     /**
      * Get a specific validation error by its error code (as defined in the ReCaptcha  configuration).
@@ -56,21 +71,6 @@ public enum ValidationError {
         }
 
         return null;
-    }
-
-    /**
-     * The error that has been thrown.
-     * Equals the error codes listed in Googles documentation.
-     */
-    private final String plainError;
-
-    /**
-     * Constructor
-     *
-     * @param plainError The original error string that has been thrown
-     */
-    ValidationError(String plainError) {
-        this.plainError = plainError;
     }
 
     /**
